+++ wikibase/data-model-services/RELEASE-NOTES.md
@@ -1,5 +1,15 @@
 # Wikibase DataModel Services release notes
 
+## Version 3.15.0 (2019-04-24)
+
+* Added `ItemLookup` implementations
+  * `LegacyAdapterItemLookup` which is an adapter to `EntityLookup`
+  * `InMemoryEntityLookup` which is a test double
+* Added `PropertyLookup` implementations
+  * `LegacyAdapterPropertyLookup` which is an adapter to `EntityLookup`
+  * `InMemoryEntityLookup` which is a test double
+* Added constructor to `InMemoryEntityLookup`
+
 ## Version 3.14.0 (2019-04-16)
 
 * Added `SeekableEntityIdPager`